add expanded option to multilingual text presenter
