+++ apps/coscrad-frontend/src/components/resources/digital-text/digital-text-page-detail-presenter.tsx
@@ -1,7 +1,7 @@
 import { IDigitalTextPage } from '@coscrad/api-interfaces';
 import { isNullOrUndefined } from '@coscrad/validation-constraints';
 import { Typography, styled } from '@mui/material';
-import { ExpandableMultilingualTextPresenter } from '../../../utils/generic-components/presenters/expandable-multilingual-text-presenter';
+import { MultilingualTextPresenter } from '../../../utils/generic-components/presenters/multilingual-text-presenter';
 import { PageContentForm, TextAndLanguage } from './page-content-form';
 
 const StyledMuiPage = styled('div')({
@@ -37,7 +37,7 @@ export const DigitalTextPageDetailPresenter = ({
      */
     return (
         <StyledMuiPage data-testid={`digital-text.page:${identifier}`}>
-            {hasContent ? <ExpandableMultilingualTextPresenter text={content} /> : null}
+            {hasContent ? <MultilingualTextPresenter text={content} expanded={false} /> : null}
             <Typography sx={{ bottom: 0, right: 0, mb: 1, mr: 1, position: 'absolute' }}>
                 {isSelected ? '**' : ''}
                 {identifier}
