+++ apps/coscrad-frontend/src/components/resources/media-items/media-item-detail.thumbnail.presenter.tsx
@@ -5,7 +5,7 @@ import {
 } from '@coscrad/api-interfaces';
 import { Card, Divider } from '@mui/material';
 import { buildDataAttributeForAggregateDetailComponent } from '../../../utils/generic-components/presenters/detail-views/build-data-attribute-for-aggregate-detail-component';
-import { ExpandableMultilingualTextPresenter } from '../../../utils/generic-components/presenters/expandable-multilingual-text-presenter';
+import { MultilingualTextPresenter } from '../../../utils/generic-components/presenters/multilingual-text-presenter';
 import { renderMediaLengthInSeconds } from '../utils/render-media-length-in-seconds-cell';
 
 export const MediaItemDetailThumbnailPresenter = ({
@@ -20,7 +20,7 @@ export const MediaItemDetailThumbnailPresenter = ({
         >
             <Card className="detail-card">
                 <div id="detail-term" className="detail-meta">
-                    <ExpandableMultilingualTextPresenter text={name} />
+                    <MultilingualTextPresenter text={name} expanded={false} />
                 </div>
                 <Divider id="detail-divider" />
                 <div className="detail-meta">
