+++ apps/coscrad-frontend/src/components/resources/playlists/episode-presenter.tsx
@@ -1,6 +1,6 @@
 import { IPlaylistEpisode } from '@coscrad/api-interfaces';
 import { Typography } from '@mui/material';
-import { ExpandableMultilingualTextPresenter } from '../../../utils/generic-components/presenters/expandable-multilingual-text-presenter';
+import { MultilingualTextPresenter } from '../../../utils/generic-components/presenters/multilingual-text-presenter';
 
 interface EpisodePresenterProps {
     episode: IPlaylistEpisode;
@@ -9,7 +9,7 @@ interface EpisodePresenterProps {
 export const EpisodePresenter = ({ episode: { name } }: EpisodePresenterProps): JSX.Element => {
     return (
         <Typography variant={'h4'}>
-            {<ExpandableMultilingualTextPresenter text={name} />}
+            {<MultilingualTextPresenter text={name} expanded={false} />}
         </Typography>
     );
 };
